1/18/2023
Do you see, gods of Olympus?! Do you need more proof than THIS?!

A choice from the gods is as useless as the gods themselves.

Turn back to Olympus, beast! I must face Zeus!

You defy the god of war?!

    ZEUS, YOUR SON HAS RETURNED! I BRING THE DESTRUCTION OF OLYMPUS!

If all of Olympus will deny me my vengeance... then all of Olympus will die. I have lived in the shadow of the gods for long enough! The time of the gods has come to an end!

ll of Olympus tremble in my name, Zeus is weak, Ares and Athena are dead and I wield the blade. We can win the Great War but not in this time! [the Great War in the background] Together we will destroy the petty gods and we will see Olympus CRUMBLE be for us! Come with me Gaia! Return to my time! Victory awaits!